+++ examples/tx/main.go
@@ -270,10 +270,8 @@ func makeOp(c *rpc.Client, t, data string) (codec.Operation, error) {
 		o = new(codec.SmartRollupExecuteOutboxMessage)
 	case tezos.OpTypeSmartRollupRecoverBond:
 		o = new(codec.SmartRollupRecoverBond)
-	case tezos.OpTypeDalAttestation:
-		o = new(codec.DalAttestation)
-	case tezos.OpTypeDalPublishSlotHeader:
-		o = new(codec.DalPublishSlotHeader)
+	case tezos.OpTypeDalPublishCommitment:
+		o = new(codec.DalPublishCommitment)
 	default:
 		return nil, fmt.Errorf("Unsupported op type %q", t)
 	}