+++ rpc/dal.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2023 Blockwatch Data Inc.
+// Copyright (c) 2024 Blockwatch Data Inc.
 // Author: alex@blockwatch.cc
 
 package rpc
@@ -9,23 +9,23 @@ import (
 
 // Ensure DAL types implement the TypedOperation interface.
 var (
-	_ TypedOperation = (*DalPublishSlotHeader)(nil)
-	_ TypedOperation = (*DalAttestation)(nil)
+	_ TypedOperation = (*DalPublishCommitment)(nil)
 )
 
-type DalPublishSlotHeader struct {
+type DalPublishCommitment struct {
 	Manager
 	SlotHeader struct {
-		Level      int64          `json:"level"`
-		Index      byte           `json:"index"`
+		Index      byte           `json:"slot_index"`
 		Commitment string         `json:"commitment"`
 		Proof      tezos.HexBytes `json:"commitment_proof"`
 	} `json:"slot_header"`
 }
 
-type DalAttestation struct {
-	Generic
-	Attestor    tezos.Address `json:"attestor"`
-	Attestation tezos.Z       `json:"attestation"`
-	Level       int64         `json:"level"`
+type DalResult struct {
+	SlotHeader struct {
+		Version    string `json:"version"`
+		Level      int64  `json:"level"`
+		Index      byte   `json:"index"`
+		Commitment string `json:"commitment"`
+	} `json:"slot_header"`
 }