+++ rpc/endorsement.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2020-2022 Blockwatch Data Inc.
+// Copyright (c) 2020-2024 Blockwatch Data Inc.
 // Author: alex@blockwatch.cc
 
 package rpc
@@ -13,11 +13,12 @@ var _ TypedOperation = (*Endorsement)(nil)
 // Endorsement represents an endorsement operation
 type Endorsement struct {
 	Generic
-	Level       int64               `json:"level"`                 // <= v008, v012+
-	Endorsement *InlinedEndorsement `json:"endorsement,omitempty"` // v009+
-	Slot        int                 `json:"slot"`                  // v009+
-	Round       int                 `json:"round"`                 // v012+
-	PayloadHash tezos.PayloadHash   `json:"block_payload_hash"`    // v012+
+	Level          int64               `json:"level"`                 // <= v008, v012+
+	Endorsement    *InlinedEndorsement `json:"endorsement,omitempty"` // v009+
+	Slot           int                 `json:"slot"`                  // v009+
+	Round          int                 `json:"round"`                 // v012+
+	PayloadHash    tezos.PayloadHash   `json:"block_payload_hash"`    // v012+
+	DalAttestation tezos.Z             `json:"dal_attestation"`       // v019+
 }
 
 func (e Endorsement) GetLevel() int64 {