+++ tezos/pvm.go
@@ -10,9 +10,13 @@ import (
 type PvmKind byte
 
 const (
-	PvmKindArith   PvmKind = iota // 0
-	PvmKindWasm200                // 1
-	PvmKindInvalid = 255
+	PvmKindArith     PvmKind = iota // 0
+	PvmKindWasm200                  // 1
+	PvmKindWasm200r1                // 2 v1
+	PvmKindWasm200r2                // 3 v2
+	PvmKindWasm200r3                // 4 v3
+	PvmKindWasm200r4                // 5 v4
+	PvmKindInvalid   = 255
 )
 
 func (t PvmKind) IsValid() bool {
@@ -38,6 +42,14 @@ func ParsePvmKind(s string) PvmKind {
 		return PvmKindArith
 	case "wasm_2_0_0":
 		return PvmKindWasm200
+	case "wasm_2_0_0.r1":
+		return PvmKindWasm200r1
+	case "wasm_2_0_0.r2":
+		return PvmKindWasm200r2
+	case "wasm_2_0_0.r3":
+		return PvmKindWasm200r3
+	case "wasm_2_0_0.r4":
+		return PvmKindWasm200r4
 	default:
 		return PvmKindInvalid
 	}
@@ -49,6 +61,14 @@ func (t PvmKind) String() string {
 		return "arith"
 	case PvmKindWasm200:
 		return "wasm_2_0_0"
+	case PvmKindWasm200r1:
+		return "wasm_2_0_0.r1"
+	case PvmKindWasm200r2:
+		return "wasm_2_0_0.r2"
+	case PvmKindWasm200r3:
+		return "wasm_2_0_0.r3"
+	case PvmKindWasm200r4:
+		return "wasm_2_0_0.r4"
 	default:
 		return ""
 	}